app[REFACT] moved editUserAccount in VM
