+++ packages/app/src/hooks/index.ts
@@ -3,3 +3,4 @@ export * from "./useTextInput";
 export * as signInScreenVM from "./signInScreenVM";
 export * as signUpScreenVM from "./signUpScreenVM";
 export * as passwordScreenVM from "./passwordScreenVM";
+export * as loggedScreenVM from "./loggedScreenVM";
